backend/logging-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
PyJWT==2.8.0
tenacity==8.2.3
httpx==0.25.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9